Track source of all packets and don't loop back.
There's a dangerous case when forwarding to a tap link where a routing
loop can occur and a packet keeps bouncing back and forward. To prevent
this, only allow writes through a Tap or a Node, track the source of
every packet and never forward packets back to the source from which
they were written.