Add support for Ethernet framing types.
It turns out that there are four different incompatible ways of
representing IPX packets on the wire in Ethernet and different hosts
need to use the same framing format to be able to communicate. Add
support for all four and default to 802.2, which is the same default
used by Novell's DOS IPX driver.