Switch stacks when processing IPX interrupts.
16-bit C code makes the assumption that ds=ss since 16-bit near pointers
don't indicate the segment. Therefore the stack must live in the same
segment as the heap and static allocations. Normally this works fine but
when servicing an interrupt the stack segment can be in an arbitrary
location. Any pointers taken to data on the stack end up becoming
invalid.

Other workarounds are to build with a different memory model or to always
use far pointers to data on the stack. However, it's probably too much
work to try to retrofit far pointers into the whole mTCP stack. The fix
taken here is to switch to a temporary stack buffer while servicing
interrupts that call into the mTCP code.