Add new implementation of TAP interface.
This reuses the code from phys/pcap.go by reimplementing the
packetDuplexStream interface to read from and write to a TAP interface.
The advantage here is that the PhysPcap code is already integrated with
the framing code, so the new TAP implementation will respect how the
user has configured to frame IPX packets.