Add implementation of ipxpkt protocol.
Some details are included in the README. For now this is just the most
basic implementation and it has not yet been fully integrated with
ipxbox. Some important features of the protocol are missing, notably
the fragmentation and reassembly code. However, it's enough for me to
successfully DHCP using mTCP and send pings.