Add passthrough non-IPX stream.
Phys by itself frames and deframes IPX packets and non-IPX packets get
dropped. This adds a NonIPX() method returning a DuplexEthernetStream
that can be used to get the filtered packets.

This is a necessary step for the ipxpkt protocol implementation since
for ipxpkt we want to wrap and forward the non-IPX Ethernet frames as
IPX packets. But we specifically don't want to ever wrap IPX in IPX -
a frame either contains an IPX packet and gets forwarded as normal, or
it contains something else and gets wrapped.