Recalculate IP/TCP/UDP checksums of captureed packets.
On modern systems the calculation of checksums is handed off to the NIC
which does it in hardware. Furthermore if we're using a virtual device like
the Linux veth* devices the checksum might not even be calculated. While
this is usually convenient as it saves CPU, it has the downside for us that
the packets that get sent out to ipxpkt.com have invalid checksums.

This can be worked around somewhat on Linux by tweaking various NIC
parameters with ethtool but I still haven't been able to get it to work
entirely reliably. So instead, we'll use gopacket to calculate the
checksums ourselves. This took some experimentation to get it to work
properly because gopacket is often too "clever" and will reconstruct some
higher-level protocols in ways which changed the packet length. So just
running gopacket.SerializePacket() doesn't work.