Change DNS resolution to happen at first connection.
DNS addresses can change IP address if dynamic DNS is being used.
Therefore change the DNS resolution so that it does not occur once
at startup, but rather just before the first time that a packet is
sent to a server. DNS resolution will be triggered again after all
active clients time out and disconnect, avoiding blocking on DNS
while a game is potentially in progress.