Correctly handle GRE key fields.
In the PPTP version of GRE, the top two octets contain the length
field while the bottom two contain the call ID. So when comparing call
ID we must mask off the top bytes. Similarly when generating the key
field to send a packet we must include the payload length.