Simplify LCP layer code.
I'm reconsidering some of the complexity in the initial design here.
This eliminates the "dialect" type in favor of just using PPPType which
is an equivalent substitute. It also removes the same from OptionType and
changes it back to just a uint8; it's not such a big deal if two options
have the same type number since they're distinguished by context in which
they're being used anyway.