Add error tracking, plus Done method.
If an error condition occurs we save an error to `n.err`. Once this
happens, stop sending or receiving any more packets. Plus add a `Done`
method that returns true if both local and remote negotiation has
succeeded, or an error has occurred.

One other minor change here is that I changed the methods that are
supposed to be called by users to upper case. Since the code is only
thread-safe if only these methods are called, this may help to avoid
mistakes.