Override receive window size to a large value.
PPTP uses a sliding window protocol as a flow control mechanism (RFC
2637, 4.2). If packets aren't acknowledged and the window becomes
full, new packets will be discarded; the window size we send here is
used as an upper limit for the window size. Since we should never
expect to reach the bandwidth limit on modern networks, we want this
to be a very large value that the client will hopefullly never reach.