Be more explicit about supported protocols.
The IPXCP negotiator is only created once we start to negotiate the
IPX parameters, but there's a race condition where we can receive an
IPXCP request from the client before we start negotiating ourself. In
that case we'll incorrectly send a Protocol-Reject for IPXCP. Instead,
work off a statically-configured set of protocols that we support.