Add internal nonblocking pipe implementation.
For routing we were using io.Pipe() to forward packets to the receive
threads. While this works okay, it has the disadvantage that writes
block until a corresponding read call, so if the reader thread is too
slow it can block processing on other threads too. Furthrmore if the
reader thread dies it can deadlock the whole system, as other clients
just need to send a packet to the dead node and they'll hang forever.

I recently got a report along these lines that I suspect was caused by
the new PPTP code and a scenario like the one above with a reader thread
dying. This new pipe implementation should avoid things like this from
happening in the future, since its Write() is nonblocking. It should
also be good for latency as well for the same reason.

In the case of eg. a broadcast packet that fans out to all other
clients, we just place the packet onto all of their receive queues and
we're done. There's no risk that we have to sequentially wait for each
reader thread to be available before we can complete and move on to the
next packet.