Add Packet type and Reader/Writer types.
Until now we've been using the io.Reader/Writer interfaces and
although they work okay, they end up being inefficient because there's
often a need to unmarshal the IPX header multiple times in different
section of code. It also makes for a somewhat clumsy API to use.

The new Reader/Writer types are analogous to the io.Reader/Writer types
but return/take an ipx.Packet, which bundles the unmarshaled IPX header
with the rest of the packet's payload. Next I'll begin converting the
existing code to use these types, and to help in this I've also defined
shim types that will help to maintain the old io APIs based on the new
ones, until they're no longer used.