Convert virtual network to use new interfaces.
The new interfaces are already allowing some code to be simplified. As
part of this, add ipx.ReadWriteCloser to the network.Node interface,
since this is (currently) the only implementation of Node in the
project. This will allow other modules that read/write to Node to start
making use of the new interfaces too.