Correctly decode Checksum and Length header fields.
This appears to have always been broken, but for a couple of reasons
we got lucky until now and it wasn't an issue. Firstly, until the
recent refactor we were passing around raw packet data, so we never
did the full roundtrip of unmarshaling the header and remarshaling
again when doing a transmit. Secondly, I don't think that DOSbox
itself pays any attention to these fields. This manifested as packets
sent over PPTP or over a bridged network connection being ignored by
the receiver.

Thanks to Nicknine on the DOSbox Deathmatch Club server for reporting
this bug.