Split virtual network into two separate networks.
There are two particular features we want here, and not necessarily both.
One is the packet forwarding part, which is akin to the behavior of a
network switch that forwards frames between Ethernet ports. The other
feature is that we usually want to filter received packets that are not
from the client's assigned IPX address, and not transmit packets back that
do not have the client's IPX address as the destination.

The way this is developing is that eventually we will be able to support
uplink connections, and in this case we don't want to do address checking.
The initial implementation of ipxswitch always broadcasts to all ports,
which is actually the behavior of an Ethernet hub. However, this is only an
internal detail, since addressable filters out any packets that aren't
appropriate for the destination client.