Separate out most dosbox-specific parts of the protocol.
This reuses much of the same approach used for network.Node, where the
protocol itself can be factored out into a wrapper that uses the client
to send and receive. Both are implementations of ipx.ReadWriteCloser, so
most of the work of running the client takes place inside a single call
to ipx.DuplexCopyPackets().

There is still a remaining TODO so this refactoring is not yet finished.