Update routing table if an address changes ports.
If an uplink client disconnects and reconnects, we may see the same
source address before from a new switch port. When this happens we don't
want to keep sending to the old switch port (which should no long exist
anyway). Instead, update the routing table to record the new
destination.