Add "fast path" check to avoid routing table changes.
Most of the time packets will not change the routing table. We also
don't need to update the lastRX field for every packet. So before
acquiring a heavy write lock for the routing table, first perform a
simple check under a read lock and return early if nothing needs to
be updated.