Increase pipe packet buffer count.
DOSbox DM club users had problems this weekend after the recent code refactor
when trying to play with four players, because the server was dropping packets
and not forwarding them. Increasing the pipe buffer size fixes the issue.
While we're here I'm taking the opportunity to rationalize the pipe API and
just always use a centrally-defined constant size rather than relying on the
caller to choose the size.

Thanks to OpenRift for reporting the bug to me.