Only capture received packets, not transmitted.
This should fix an issue discovered when testing uplink where dosbox clients
and uplink clients could not completely see each other. What I believe was
happening was the following:

1. dosbox client sends a broadcast packet of some type.
2. packet is forwarded to uplink client and transmitted via pcap
3. pcap captures and redelivers the transmitted packet.
4. uplink client sends redelivered broadcast back to server
5. server delivers packet again back to dosbox client.

For each IPX address, ipxswitch tracks the last switch port from which it
delivered a packet. This means that once (4) happens the uplink client has
"stolen" the address of the dosbox client; any packets sent to its address
will be routed to the uplink client instead. This is only possible because
uplinks are non-addressable so there's no sanity checking of source addresses
like there is with dosbox clients.

Thanks to TheBaratusII for reporting and helping me to gather data to track
down the source of this bug.