Detect and discard looped-back packets.
Bug #18 occurs because npcap re-captures packets that we have transmitted
ourselves. After lots of digging into how npcap works I've given up on
making a proper solution that fixes the root cause. Instead, we can use the
IPX TransControl field as a kind of tagging mechanism. All packets written
to physical devices have this field set to a special value, and if we
detect this value in a captured packet then we will discard it. As a bonus,
this should also fix other potential routing loop issues if eg. two uplinks
connect the same server to the same network.

Fixes #18.