Fix crash with `nil` address
This crash was introduced recently in commit 9d49efb594075052 when I
switched the logging code to call `addrAttr()` at the time that a server
connection is established. It looks like this would already have caused
a crash with the previous code, but only in an error scenario.

Thanks to @CyberXZT for reporting the bug. Fixes #43.